feat: [loki-exporter] Fixed service name and trace id and span id generation in exporters
