+++ server/ingester/flow_log/exporter/config_test.yaml
@@ -30,8 +30,12 @@ ingester:
         max-retries: 2
         static-labels:
           foo: bar
-        export-datas: [cbpf-net-span,ebpf-sys-span]
-        export-data-types: [service_info,tracing_info,network_layer,flow_info,client_universal_tag,server_universal_tag,tunnel_info,transport_layer,application_layer,capture_info,native_tag,metrics]
+        export-datas:
+          - ebpf-sys-span
+        export-data-types:
+          - service_info
+          - tracing_info
+          - network_layer
         export-only-with-traceid: false
         log-format:
           mapping:
