+++ server/ingester/flow_log/exporter/loki_exporter/fmt.go
@@ -30,7 +30,7 @@ var defaultHeader = map[string]string{
 }
 
 func (le *LokiExporter) buildLogHeader() {
-	defaultLogHeaderFormat := "{time}=%s, {service_name}=%s, {log_level}=%s, {trace_id}=%s, {span_id}=%s, "
+	defaultLogHeaderFormat := `{time}="%s",{service_name}="%s",{log_level}="%s",{trace_id}="%s",{span_id}="%s",`
 	for k, v := range defaultHeader {
 		if _, ok := le.cfg.LogFmt.Mapping[k]; ok {
 			v = le.cfg.LogFmt.Mapping[k]
@@ -54,40 +54,40 @@ func responseStatusToLogLevel(status uint8) LogLevel {
 
 func buildLogBodyDNS(l7 *log_data.L7FlowLog) string {
 	return fmt.Sprintf(
-		"request_type=%s, request_resource=%s, response_code=%d, response_exception=%s, response_result=%s",
+		`request_type="%s",request_resource="%s",response_code=%d,response_exception="%s",response_result=%s`,
 		l7.RequestType, l7.RequestResource, l7.ResponseCode, l7.ResponseException, l7.ResponseResult,
 	)
 }
 
 func buildLogBodyHTTP(l7 *log_data.L7FlowLog) string {
 	return fmt.Sprintf(
-		"method=%s, name=%s, path=%s, status_code=%d, response_exception=%s",
+		`method="%s",name="%s",path="%s",status_code=%d,response_exception=%s`,
 		l7.RequestType, l7.RequestDomain, l7.RequestResource, l7.ResponseCode, l7.ResponseException,
 	)
 }
 func buildLogBodyDubbo(l7 *log_data.L7FlowLog) string {
 	return fmt.Sprintf(
-		"rpc_system=apache_dubbo, service=%s, method=%s, request_domain=%s, dubbo_version=%s, response_code=%d, response_exception=%s",
+		`rpc_system="apache_dubbo",service="%s",method="%s",request_domain="%s",dubbo_version="%s",response_code=%d,response_exception=%s`,
 		l7.RequestResource, l7.RequestType, l7.RequestDomain, l7.Version, l7.ResponseCode, l7.ResponseException,
 	)
 }
 func buildLogBodyGRPC(l7 *log_data.L7FlowLog) string {
 	return fmt.Sprintf(
-		"rpc_system=grpc, service=%s, method=%s, request_domain=%s, http_flavor=%s, response_code=%d, response_exception=%s",
+		`rpc_system="grpc",service="%s",method="%s",request_domain="%s",http_flavor="%s",response_code=%d,response_exception=%s`,
 		l7.RequestResource, l7.RequestType, l7.RequestDomain, l7.Version, l7.ResponseCode, l7.ResponseException,
 	)
 }
 
 func buildLogBodyKafka(l7 *log_data.L7FlowLog) string {
 	return fmt.Sprintf(
-		"messaging_system=kafka, request_type=%s, request_resource=%s, request_domain=%s, response_code=%d, response_exception=%s",
+		`messaging_system="kafka",request_type="%s",request_resource="%s",request_domain="%s",response_code=%d,response_exception=%s`,
 		l7.RequestType, l7.RequestResource, l7.RequestDomain, l7.ResponseCode, l7.ResponseException,
 	)
 }
 
 func buildLogBodyMQTT(l7 *log_data.L7FlowLog) string {
 	return fmt.Sprintf(
-		"messaging_system=mqtt, request_type=%s, request_resource=%s, request_domain=%s, response_code=%d, response_exception=%s",
+		`messaging_system="mqtt",request_type="%s",request_resource="%s",request_domain="%s",response_code=%d,response_exception=%s`,
 		l7.RequestType, l7.RequestResource, l7.RequestDomain, l7.ResponseCode, l7.ResponseException,
 	)
 }
@@ -95,7 +95,7 @@ func buildLogBodyMQTT(l7 *log_data.L7FlowLog) string {
 func buildLogBodyMySQL(l7 *log_data.L7FlowLog) string {
 	_, operation := exporter_common.GetSQLSpanNameAndOperation(l7.RequestResource)
 	return fmt.Sprintf(
-		"db_system=mysql, operation=%s, statement=%s, request_type=%s, response_exception=%s",
+		`db_system="mysql",operation="%s",statement="%s",request_type="%s",response_exception=%s`,
 		operation, l7.RequestResource, l7.RequestType, l7.ResponseException,
 	)
 }
@@ -103,7 +103,7 @@ func buildLogBodyMySQL(l7 *log_data.L7FlowLog) string {
 func buildLogBodyPostgreSQL(l7 *log_data.L7FlowLog) string {
 	_, operation := exporter_common.GetSQLSpanNameAndOperation(l7.RequestResource)
 	return fmt.Sprintf(
-		"db_system=postgresql, operation=%s, statement=%s, request_type=%s, response_exception=%s",
+		`db_system="postgresql",operation="%s",statement="%s",request_type="%s",response_exception=%s`,
 		operation, l7.RequestResource, l7.RequestType, l7.ResponseException,
 	)
 }
@@ -111,7 +111,7 @@ func buildLogBodyPostgreSQL(l7 *log_data.L7FlowLog) string {
 func buildLogBodyRedis(l7 *log_data.L7FlowLog) string {
 	_, operation := exporter_common.GetSQLSpanNameAndOperation(l7.RequestResource)
 	return fmt.Sprintf(
-		"db_system=redis, operation=%s, statement=%s, request_type=%s, response_exception=%s",
+		`db_system="redis",operation="%s",statement="%s",request_type="%s",response_exception=%s`,
 		operation, l7.RequestType, l7.RequestResource, l7.ResponseException,
 	)
 }
